+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-enumerable` will be documented in this file
 
+## 1.1.0 - 2019-01-18
+
+- Update mindtwo/laravel-dynamic-model-mutators to 2.0.1
+- Changed property scope to public
+
 ## 1.0.0 - 2018-12-03
 
 - Initial release